requirements.txt
blinker==1.8.2
boto3
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pdfkit==1.0.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.19.2
